% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCAR_rgeneric_vk.R
\name{inla.rgeneric.MCAR.model_v2}
\alias{inla.rgeneric.MCAR.model_v2}
\title{A}
\usage{
inla.rgeneric.MCAR.model_v2(
  cmd = c("graph", "Q", "mu", "initial", "log.norm.const", "log.prior", "quit"),
  theta = NULL
)
}
\value{
B
}
\description{
A
}
\examples{
1 + 1
}
